requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
Pillow==6.2.1
requests==2.22.0
stdeb==0.9.0
urllib3==1.25.7
